Change timer to 8-bit

+++ src/timer.rs
@@ -1,10 +1,10 @@
 pub struct Timer {
-    pub current: u16,
+    pub current: u8,
     pub active: bool
 }
 
 impl Timer {
-    pub fn new(initial: u16) -> Timer {
+    pub fn new(initial: u8) -> Timer {
         Timer {
             current: initial,
             active: true
